Reminder repository Updated

+++ SmartAssistant.Shared/Repositories/ReminderRepository.cs
@@ -24,26 +24,31 @@ namespace SmartAssistant.Shared.Repositories
 			await context.SaveChangesAsync();
 		}
 
-		public async System.Threading.Tasks.Task DeleteAsync(ReminderModel entity)
-		{
-			var reminder = mapper.Map<Reminder>(entity);
-			context.Reminders.Remove(reminder);
-			await context.SaveChangesAsync();
-		}
+        public async System.Threading.Tasks.Task DeleteAsync(ReminderModel entity)
+        {
+            var reminder = await context.Reminders.FindAsync(entity.Id);
+            if (reminder != null)
+            {
+                context.Reminders.Remove(reminder);
+                await context.SaveChangesAsync();
+            }
+        }
 
-		public async System.Threading.Tasks.Task<IEnumerable<ReminderModel>> GetAllAsync()
+
+        public async System.Threading.Tasks.Task<IEnumerable<ReminderModel>> GetAllAsync()
 		{
 			var reminder = await context.Reminders.ToListAsync();
 			return mapper.Map<List<ReminderModel>>(reminder);
 		}
 
-		public async System.Threading.Tasks.Task<ReminderModel> GetByIdAsync(int id)
-		{
-			var reminder = await context.Reminders.FindAsync(id);
-			return mapper.Map<ReminderModel>(reminder);
-		}
+        public async System.Threading.Tasks.Task<ReminderModel> GetByIdAsync(int id)
+        {
+            var reminder = await context.Reminders.AsNoTracking().FirstOrDefaultAsync(r => r.Id == id);
+            return mapper.Map<ReminderModel>(reminder);
+        }
 
-		public async System.Threading.Tasks.Task<List<ReminderModel>> GetRemindersByUserIdAsync(string userId)
+
+        public async System.Threading.Tasks.Task<List<ReminderModel>> GetRemindersByUserIdAsync(string userId)
 		{
 			var reminder = await context.Reminders.Where(x => x.UserId == userId).ToListAsync();
 			return mapper.Map<List<ReminderModel>>(reminder);
